Refactor resetting
