+++ conda_self/cli/main_reset.py
@@ -14,9 +14,11 @@ def configure_parser(parser: argparse.ArgumentParser) -> None:
 
 
 def execute(args: argparse.Namespace) -> int:
-    from .common import reset
+    from ..reset import reset
+    from ..query import permanent_dependencies
 
     print("Resetting 'base' environment...")
-    reset()
+    uninstallable_packages = permanent_dependencies()
+    reset(uninstallable_packages=uninstallable_packages)
 
     return 0
