+++ conda_self/reset.py
@@ -3,16 +3,15 @@ from conda.core.prefix_data import PrefixData
 from conda.core.link import PrefixSetup, UnlinkLinkTransaction
 from conda.base.context import context
 
-from ..query import permanent_dependencies
+from .query import permanent_dependencies
 
 
-def reset():
-    installed = sorted(PrefixData(sys.prefix).iter_records(), key=lambda x: x.name)
-    uninstallable_packages = permanent_dependencies()
+def reset(prefix: str = sys.prefix, uninstallable_packages: list[str] = []):
+    installed = sorted(PrefixData(prefix).iter_records(), key=lambda x: x.name)
     packages_to_remove = [pkg for pkg in installed if pkg.name not in uninstallable_packages]
 
     stp = PrefixSetup(
-        target_prefix=sys.prefix,
+        target_prefix=prefix,
         unlink_precs=packages_to_remove,
         link_precs=(),
         remove_specs=(),
